Second Coomit

+++ app/src/main/java/com/example/androidbootcamp2021_atul/excersise2/Q2.kt
@@ -36,7 +36,7 @@ class Q2{
         println("Sum of Int = ${x+y}")
     }
 
-    fun Operations(x: Double, y: Double) {
+    fun Operations(x: Long, y: Long) {
         println("Sum of Double = ${x+y}")
     }
